Organize shared sensor benchmarks
Move sensor-specific CLI, timing, reporting, and ray-caster helpers into a coherent package shared by all physics backends. Replace the redundant cross-product subprocess checks with direct parser coverage while retaining backend-specific ray-caster validation.
