+++ source/isaaclab/isaaclab/benchmark/sensor_suites/__init__.py
@@ -1,0 +1,18 @@
+# Copyright (c) 2022-2026, The Isaac Lab Project Developers (https://github.com/isaac-sim/IsaacLab/blob/main/CONTRIBUTORS.md).
+# All rights reserved.
+#
+# SPDX-License-Identifier: BSD-3-Clause
+
+"""Shared infrastructure for backend sensor micro-benchmarks."""
+
+from .cli import add_sensor_benchmark_args
+from .ray_caster import rough_terrain_size
+from .timing import SensorLatencySamples, add_sensor_latency_measurements, collect_sensor_latency_samples
+
+__all__ = [
+    "SensorLatencySamples",
+    "add_sensor_benchmark_args",
+    "add_sensor_latency_measurements",
+    "collect_sensor_latency_samples",
+    "rough_terrain_size",
+]
