+++ source/isaaclab/test/benchmark/test_cli.py
@@ -9,8 +9,8 @@ import argparse
 
 import pytest
 
-import isaaclab.benchmark._cli as benchmark_cli
 from isaaclab.benchmark._cli import parse_non_negative_int, parse_positive_int, validate_warmup_steps
+from isaaclab.benchmark.sensor_suites import add_sensor_benchmark_args
 
 
 @pytest.mark.parametrize("argument_type", [parse_non_negative_int, parse_positive_int])
@@ -38,7 +38,7 @@ def test_validate_warmup_steps_rejects_exhausted_workload(warmup_steps: int, ava
 def test_sensor_argument_helper_registers_one_consistent_common_contract() -> None:
     """Every retained sensor script should share the same workload and output arguments."""
     parser = argparse.ArgumentParser()
-    benchmark_cli.add_sensor_benchmark_args(
+    add_sensor_benchmark_args(
         parser,
         physics_variants=("newton_mjwarp", "newton_kamino"),
         default_physics_variant="newton_mjwarp",
@@ -76,3 +76,28 @@ def test_sensor_argument_helper_registers_one_consistent_common_contract() -> No
         "benchmark_formatter": "json",
         "device": "cpu",
     }
+
+
+@pytest.mark.parametrize(
+    ("args", "message"),
+    [
+        (("--num_envs", "0"), "must be greater than zero"),
+        (("--num_steps", "0"), "must be greater than zero"),
+        (("--warmup_steps", "-1"), "must be non-negative"),
+        (("--physics_variant", "unknown"), "invalid choice"),
+    ],
+)
+def test_sensor_argument_helper_rejects_invalid_common_arguments(args: tuple[str, str], message: str, capsys) -> None:
+    """Invalid common sensor arguments should fail through argparse."""
+    parser = argparse.ArgumentParser()
+    add_sensor_benchmark_args(
+        parser,
+        physics_variants=("newton_mjwarp", "newton_kamino"),
+        default_physics_variant="newton_mjwarp",
+        add_device=False,
+    )
+
+    with pytest.raises(SystemExit):
+        parser.parse_args(args)
+
+    assert message in capsys.readouterr().err
