+++ source/isaaclab_ovphysx/benchmark/sensors/benchmark_joint_wrench.py
@@ -18,7 +18,7 @@ from __future__ import annotations
 import argparse
 from functools import partial
 
-from isaaclab.benchmark._cli import add_sensor_benchmark_args
+from isaaclab.benchmark.sensor_suites import add_sensor_benchmark_args
 
 parser = argparse.ArgumentParser(description="Benchmark the OVPhysX JointWrench sensor update path.")
 add_sensor_benchmark_args(
@@ -36,7 +36,7 @@ from isaaclab_ovphysx.physics import OvPhysxCfg
 
 import isaaclab.sim as sim_utils  # noqa: F401
 from isaaclab.benchmark import LatencyBenchmarkRunner, SingleMeasurement
-from isaaclab.benchmark.micro import add_sensor_latency_measurements, collect_sensor_latency_samples
+from isaaclab.benchmark.sensor_suites import add_sensor_latency_measurements, collect_sensor_latency_samples
 from isaaclab.scene import InteractiveScene, InteractiveSceneCfg
 from isaaclab.sensors import JointWrenchSensorCfg
 from isaaclab.sim import SimulationCfg, build_simulation_context
