+++ source/isaaclab_ovphysx/benchmark/sensors/benchmark_ray_caster.py
@@ -19,8 +19,7 @@ from __future__ import annotations
 import argparse
 from functools import partial
 
-from isaaclab.benchmark._cli import add_sensor_benchmark_args
-from isaaclab.benchmark._ray_caster import rough_terrain_size
+from isaaclab.benchmark.sensor_suites import add_sensor_benchmark_args, rough_terrain_size
 
 parser = argparse.ArgumentParser(description="Benchmark the OVPhysX RayCaster update path.")
 add_sensor_benchmark_args(
@@ -50,7 +49,7 @@ from isaaclab_ovphysx.physics import OvPhysxCfg
 import isaaclab.sim as sim_utils
 from isaaclab.assets import AssetBaseCfg, RigidObjectCfg
 from isaaclab.benchmark import LatencyBenchmarkRunner, SingleMeasurement
-from isaaclab.benchmark.micro import add_sensor_latency_measurements, collect_sensor_latency_samples
+from isaaclab.benchmark.sensor_suites import add_sensor_latency_measurements, collect_sensor_latency_samples
 from isaaclab.scene import InteractiveScene, InteractiveSceneCfg
 from isaaclab.sensors import RayCasterCfg, patterns
 from isaaclab.sim import SimulationCfg, build_simulation_context
